Extract implementation of rings into two separate files
